+++ viewer/SelfTest.lua
@@ -107,7 +107,7 @@ function love.load()
   }
   local modChunk = assert(love.filesystem.load("main.lua"))
   modChunk(mod)
-  assert(mod.exports.version == "0.5.0", "wrong mod export version")
+  assert(mod.exports.version == "0.5.1", "wrong mod export version")
   assert(type(handlers["battle.started"]) == "function",
          "battle integration event was not registered")
   local liveInner = {}